pyproject.toml
[project]
name = "llmcourse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "afinn>=0.1",
    "dspy>=3.1.3",
    "fastapi>=0.128.5",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "nltk>=3.9.2",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "qlever>=0.5.45",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.3.0",
    "spacy>=3.8.11",
    "uvicorn>=0.40.0",
]
